feat: refactoring code / add readme and glossary support / add support for glue and redshift connectors
