+++ atlanapi/delete_asset.py
@@ -1,21 +1,22 @@
 import logging
-import json
 
 from atlanapi.ApiConfig import create_api_config
 from atlanapi.atlanutils import AtlanApiRequest
 
-api_conf = create_api_config()
-search_headers = {
-    'Content-Type': 'application/json;charset=utf-8',
-    'Authorization': api_conf.api_token
-}
-
 logger = logging.getLogger('main_logger')
+
+api_conf = create_api_config()
+authorization = 'Bearer {}'.format(api_conf.api_token)
+
+headers = {
+    'Authorization': authorization,
+    'Content-Type': 'application/json'
+}
 
 
 def delete_asset(asset_guid):
-    search_url = 'https://{}/api/meta/entity/bulk?guid={}&deleteType=HARD#bulk_delete_assets"'.format(api_conf.instance, asset_guid)
-    atlan_api_query_request_object = AtlanApiRequest("DELETE", search_url, search_headers, {})
+    delete_url = 'https://{}/api/meta/entity/bulk?guid={}&deleteType=HARD"'.format(api_conf.instance, asset_guid)
+    atlan_api_query_request_object = AtlanApiRequest("DELETE", delete_url, headers, {})
     try:
         return atlan_api_query_request_object.send_atlan_request().status_code == 200
     except Exception as e:
