+++ atlanapi/get_entity_columns.py
@@ -20,7 +20,7 @@ def get_entity_columns(entity_guid):
     atlan_api_query_request_object = AtlanApiRequest("GET", search_url, search_headers, {})
     try:
         columns_response = json.loads(atlan_api_query_request_object.send_atlan_request().text)
-        columns = columns_response['entity']['attributes']['columns']
+        columns = columns_response['entity']['relationshipAttributes']['columns']
         if columns:
             return {column['displayText']: column['guid'] for column in columns}
         # Not really an error, because it is normal that an entity doesn't have columns at first.
