+++ atlanapi/linkTerm.py
@@ -3,35 +3,53 @@ import logging
 
 from atlanapi.ApiConfig import create_api_config
 from atlanapi.atlanutils import AtlanApiRequest
-from atlanapi.searchAssets import get_asset_guid_by_qualified_name
 from atlanapi.searchGlossaryTerms import get_glossary_term_guid_by_name
-from model.Asset import EntityLineage, ColumnLineage
+from model import TableLineage, ColumnLineage
 
 logger = logging.getLogger('main_logger')
 
+api_conf = create_api_config()
+
+authorization = 'Bearer {}'.format(api_conf.api_token)
+headers = {
+    'Authorization': authorization,
+    'Content-Type': 'application/json'
+}
+
 
 def link_term(asset):
-    if isinstance(asset, ColumnLineage) or isinstance(asset, EntityLineage) or not asset.term or not asset.glossary:
+    if isinstance(asset, ColumnLineage) or isinstance(asset, TableLineage) or not asset.term or not asset.glossary:
         return
     try:
         term_guid = get_glossary_term_guid_by_name(asset.term, asset.glossary)
-        asset_qualified_name = asset.get_qualified_name()
-        asset_info = get_asset_guid_by_qualified_name(asset_qualified_name, asset.get_atlan_type_name())
-        asset_guid = asset_info['guid']
-        payload = json.dumps([{"guid": asset_guid}])
-        api_conf = create_api_config()
-        headers = {
-            'Content-Type': 'application/json;charset=utf-8',
-            'Authorization': api_conf.api_token
-        }
-        link_to_term_url = 'https://{}/api/metadata/atlas/tenants/default/glossary/terms/{}/assignedEntities'.format(
-            api_conf.instance, term_guid)
-
+        payload = json.dumps({
+            "entities":
+                [
+                    {
+                        "typeName": asset.get_atlan_type_name(),
+                        "attributes": {
+                            "name": asset.get_asset_name(),
+                            "qualifiedName": asset.get_qualified_name()
+                        },
+                        "relationshipAttributes": {
+                            "meanings": [
+                                {
+                                    "typeName": "AtlasGlossaryTerm",
+                                    "guid": term_guid
+                                }
+                            ]
+                        }
+                    }
+                ]
+            })
+        link_to_term_url = 'https://{}/api/meta/entity/bulk#{}'.format(api_conf.instance, 'attachGlossaryTerm')
         atlan_api_request_object = AtlanApiRequest("POST", link_to_term_url, headers, payload)
 
         atlan_api_request_object.send_atlan_request()
-        logger.debug("Glossary term '{}' linked successfully to {} '{}'"
-                     .format(asset.term, asset.get_atlan_type_name(), asset.get_asset_name()))
+        logger.info("Glossary term '{}' linked successfully to {} '{}'".format(
+            asset.term, asset.get_atlan_type_name(), asset.get_asset_name())
+        )
+
     except Exception as e:
         logger.warning("Error while linking glossary term '{}' to {} '{}'. Glossary term must already exist in Atlan."
                        .format(asset.term, asset.get_atlan_type_name(), asset.get_asset_name()))
